double

+++ src/query/constants.hpp
@@ -2,10 +2,10 @@
 
 #include <cstddef>
 #include <cstdint>
-#include <string.hpp>
 #include <pair.hpp>
+#include <string.hpp>
 
 namespace fast::query {
-  constexpr size_t MAX_RESULTS = 10;
-  typedef pair<uint32_t, string> Result;
-}
+constexpr size_t MAX_RESULTS = 10;
+typedef pair<double, string> Result;
+}  // namespace fast::query
